+++ selene/img/typed/ImageView.hpp
@@ -571,7 +571,7 @@ bool equal(const ImageView<PixelType0, modifiability_0>& img_0, const ImageView<
     // std::equal may not be optimized to std::memcmp, even though we're dealing with a POD-type here...
     // const bool equal_row = std::equal(begin0, end0, begin1);
     // ...so let's just call std::memcmp directly:
-    const auto nr_bytes = std::distance(begin0, end0) * sizeof(PixelType0);
+    const auto nr_bytes = static_cast<std::size_t>(std::distance(begin0, end0)) * sizeof(PixelType0);
     const bool equal_row = (std::memcmp(begin0, begin1, nr_bytes) == 0);
 
     if (!equal_row)
